tests: Allow running tests against Unity's bundled mono
DebugTests now look at UNITY_PATH environment variable to locate a
version of mono that is bundled with a real Unity install to run its
tests.  It is currently hard-coded to use the 4.7.1-api environment.

If UNITY_PATH is not given, it attempts to use MONO_PATH to locate the
exe instead.  If neither available, an exeception is thrown.