Add attach handler that doesn't rely on process detection
It is not strictly necessary to use Unity process detection for
attaching as editor or player PID can be known in other ways. Detection
also does not work for players running on, for example, Android
emulator.

This patch adds a mechanism to skip UnityDebug's process detection
mechanism, with the aim for it to become the default once sufficiently
mature.  For now, it is activated by specifying "projectRoot",
"platform", or "port" as part of the debug configuration.

Corresponding configuration schemas are also added.