Add Android attach support
This adds attaching to Android players running both in emulator and
on-device.

When platform is set as 'android' and an explicit port is not provided,
the adapter will attempt to autodetect Unity players it can connect to
using adb. The devices to search for can be narrowed futher by setting
'android.connection' or ANDROID_SERIAL:

    // Connect to USB connected device
    {
       "type": "unity",
       "platform": "android",
       ...
       "android": { "connection": "usb" }
    }

    // Connect to first emulator instance
    {
       "type": "unity",
       "platform": "android",
       ...
       "env": {
           "ANDROID_SERIAL": "emulator-5554"
       }
    }

Note that player detection is not 100% reliable as there is no reliable
way to map open ports to processes on vanilla Android. The detection
algorithm assumse any open ports in the expected range is a player
instance.